refactor: move reputationSchema to the shared/model
