+++ packages/apps/dashboard/client/src/features/leaderboard/model/leaderboardSchema.ts
@@ -1,19 +1,6 @@
 import { z } from 'zod';
 
-export const reputationSchema = z.unknown().transform((value) => {
-  try {
-    const knownReputation = z
-      .union([z.literal('low'), z.literal('medium'), z.literal('high')])
-      .parse(value);
-
-    return knownReputation;
-  } catch (error) {
-    console.error(error);
-    return 'unknown';
-  }
-});
-
-export type Reputation = z.infer<typeof reputationSchema>;
+import { reputationSchema } from '@/shared/model/reputationSchema';
 
 const leaderboardEntity = z.object({
   address: z.string(),
