+++ packages/apps/dashboard/client/src/features/searchResults/model/addressDetailsSchema.ts
@@ -1,7 +1,7 @@
 import { Role } from '@human-protocol/sdk/src/constants';
 import { z } from 'zod';
 
-import { reputationSchema } from '@/features/leaderboard/model/leaderboardSchema';
+import { reputationSchema } from '@/shared/model/reputationSchema';
 
 const transformOptionalTokenAmount = (
   value: string | undefined | null,
